Late night bug fixes
It was a late when I committed the constants and interfaces of ElectronicValueRW, and in some places, UPOS specification specifies wrong version numbers. This resulted in some wrong or missing version comments, senseless constant values and duplicate and missing method declarations in ...Control115, ...Service115 and ...Const interfaces.
Has been corrected.