Meaningless Constants Removal
Removal of GraphicDisplay constant GDSP_ST_HOST_HARDTOTALS and SoundPlayer constant SPLY_ST_HOST_HARDTOTALS. Both constants are meaningless because they should be values of the Storage property of the corresponding device which specifies the source for loading or playing data. Therefore, Storage must specify the source as host or HardTotals, not both.